* Big Change*
Removed list object from atom class. For the moment, keeping the list attribute as a Boolean.
Moved all the logic of adding and removing atoms to lattice classes.
Removed extract, deposit, addOwnAtom and addBond Atom from atoms.
Removed addTotalProbability method from atom.
Changed name attribute from totalProbability to probability for atoms.
Redesign of ModifiedBuffer, removing intermediate private functions.
Renamed updateAllRates to updateRate and obtainRatesFromNeighbour to obtainRateFromNeighbours.
In the case of binned list, probability is only changed from parent. So, created a method setTotalProbability in AbstractList and overriding addToTotalProbability.

* Note *

Binned list is considerably slower, this is due to the fact that the total probability is not updated correctly and instead is calculated every time from the list of atoms.